refactor(repositories): implement repository pattern for colleges, courses, degrees, departments, and sections with pagination support
