+++ lib/db/repositories/college_repository.ts
@@ -1,0 +1,25 @@
+import connect from "../db";
+import { CollegeSchema } from "../db_types";
+import College from "../models/colleges";
+
+/**
+ * Retrieves a single department by its ID.
+ * @param {number} id - The ID of the department.
+ * @returns {Promise<DepartmentSchema | null>} The department object, or null if not found.
+ */
+export const getCollege = async (id: number): Promise<CollegeSchema | null> => {
+  await connect();
+  const college = await College.findById(id).lean<CollegeSchema>();
+  return college;
+};
+
+
+/**
+ * Retrieves all colleges stored in the database.
+ * * @returns {Promise<CollegeSchema[]>} An array of all colleges.
+ */
+export const getColleges = async (): Promise<CollegeSchema[]> => {
+  await connect();
+  const colleges = await College.find({}).lean<CollegeSchema[]>();
+  return colleges;
+};
